Update README.md
This was a project developed by a team of 4 students (myself, John Jackson, Khoi Le, and James Nance) for the class TCSS360 at University of Washington Tacoma.

The application manages paper submission, storage, retrieval, and review for an arbitrary number of distinct conferences, based on user accounts and a set of potential user roles in particular conferences.  The storage mechanism was an SQLite database.

I was responsible for much of the design and the majority of the gui implementation.  All classes are appropriately commented with authorship information.  The larger part of this particular exercise was in approaches to planning and execution of a project through stages of a product lifecycle.